application/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["tobias <tobias.lindert@factor-a.com>"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
google-cloud-storage = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
